+++ lib/rspc.dart
@@ -19,4 +19,4 @@ part "src/rspc/compiler.dart";
 part "src/rspc/tag.dart";
 part "src/rspc/tag_util.dart";
 
-const VERSION = "0.7.4";
+const VERSION = "0.7.5";