+++ lib/src/server.dart
@@ -118,6 +118,17 @@ abstract class StreamServer {
    */
   int sessionTimeout;
 
+  /** Whether the response is default to the chunked transfer encoding
+   * mode. If true, RSP pages and static resources will be gzipped if
+   * the client accepts it.
+   *
+   * Default: true.
+   *
+   * You can turn it off if you prefer not to compress the output.
+   * For exaple, it might be compressed by a front-end server.
+   */
+  bool chunkedTransferEncoding;
+
   /** Indicates whether the server is running.
    */
   bool get isRunning;