+++ lib/src/server_impl.dart
@@ -7,7 +7,7 @@ part of stream;
 typedef void _ConnectErrorCallback(HttpConnect connect, err, [stackTrace]);
 
 class _StreamServer implements StreamServer {
-  final String version = "0.7.4";
+  final String version = "0.7.5";
   HttpServer _server;
   var _host = InternetAddress.ANY_IP_V4;
   int _port = 8080;
@@ -196,6 +196,9 @@ class _StreamServer implements StreamServer {
     if (_server != null)
       _server.sessionTimeout = _sessTimeout;
   }
+
+  @override
+  bool chunkedTransferEncoding = true;
 
   @override
   ResourceLoader get resourceLoader => _resLoader;