+++ test/syntax/Issue4.rsp.dart
@@ -11,9 +11,7 @@ Future Issue4(HttpConnect connect) { //#1
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-
-  if (!connect.isIncluded)
-    response.headers.contentType = ContentType.parse("text/html; charset=utf-8");
+  Rsp.init(connect, "text/html; charset=utf-8");
 var foo = "test";
 render(HttpConnect connect, {more, less}) {
 }