+++ test/syntax/include.rsp.dart
@@ -11,9 +11,7 @@ Future include(HttpConnect connect, {foo, more, less}) { //#3
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-
-  if (!connect.isIncluded)
-    response.headers.contentType = ContentType.parse("text/html; charset=utf-8");
+  Rsp.init(connect, "text/html; charset=utf-8");
 
   var less = new StringBuffer(); _cs_.add(connect); //var#3
   connect = new HttpConnect.stringBuffer(connect, less); response = connect.response;