+++ test/syntax/lastModified1.rsp.dart
@@ -11,10 +11,8 @@ Future lastModified1(HttpConnect connect, {input}) { //#2
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-
-  if (!connect.isIncluded)
-    response.headers.contentType = ContentType.parse("text/html; charset=utf-8");
-  response.headers.set(HttpHeaders.LAST_MODIFIED, new DateTime.fromMillisecondsSinceEpoch(1371817597082));
+  Rsp.init(connect, "text/html; charset=utf-8",
+    () => new DateTime.fromMillisecondsSinceEpoch(1371822684138));
 
   response.write("""<html>
   <head>