+++ CHANGES.md
@@ -3,6 +3,7 @@
 **0.7.4**
 
 * Issue 36: Change the spec of [:json] to generate a JSON object that can be parse as Dart object directly
+* Issue 35: A tag for generating the dart script for development, while generating only JS in production
 
 **0.7.3**